fix:transaction-applied-v3

See merge request s13-fintech-finance-sub1/S13P21A407!177

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/mission/dto/request/MissionRedisDto.java
@@ -21,5 +21,5 @@ public class MissionRedisDto implements Serializable {
     private LocalDateTime validFrom;
     private LocalDateTime validTo;
     private Status status = Status.IN_PROGRESS;
-    private Integer progress;
+    private Long progress;
 }
